added backend support for anthropic claude via aws bedrock
