+++ packages/flowtest-electron/package.json
@@ -27,6 +27,15 @@
   },
   "dependencies": {
     "@apidevtools/swagger-parser": "^10.1.0",
+    "@aws-crypto/sha256-js": "^5.2.0",
+    "@aws-sdk/client-bedrock": "^3.583.0",
+    "@aws-sdk/client-bedrock-runtime": "^3.583.0",
+    "@aws-sdk/credential-provider-node": "^3.583.0",
+    "@aws-sdk/types": "^3.577.0",
+    "@smithy/eventstream-codec": "^3.0.0",
+    "@smithy/protocol-http": "^4.0.0",
+    "@smithy/signature-v4": "^3.0.0",
+    "@smithy/util-utf8": "^3.0.0",
     "axios": "^1.6.7",
     "chokidar": "^3.6.0",
     "dotenv": "^16.4.5",
