+++ src/components/molecules/flow/flowtestai.js
@@ -7,16 +7,19 @@ const translateGeneratedNodesToOpenApiNodes = (generatedNodes, openApiNodes) => 
   generatedNodes.forEach((gnode, _) => {
     const node = openApiNodes.find((node) => node.operationId === gnode.name);
     if (node !== undefined) {
-      const outputNode = { ...node };
+      // take best of both worlds (parameters and request body generated by ai model and ones intelligently generated by parser)
+      const outputNode = JSON.parse(JSON.stringify(node)); //{ ...node };
       const node_arguments = JSON.parse(gnode.arguments);
-      if (node_arguments.requestBody) {
-        outputNode['requestBody'] = {};
-        outputNode['requestBody']['type'] = 'raw-json';
-        outputNode['requestBody']['body'] = JSON.stringify(node_arguments.requestBody);
-      }
+      // if (node_arguments.requestBody) {
+      //   outputNode['requestBody'] = {};
+      //   outputNode['requestBody']['type'] = 'raw-json';
+      //   outputNode['requestBody']['body'] = JSON.stringify(node_arguments.requestBody);
+      // }
       if (node_arguments.parameters) {
-        outputNode.preReqVars = {};
         Object.entries(node_arguments.parameters).forEach(([paramName, paramValue], _) => {
+          if (!outputNode.preReqVars) {
+            outputNode.preReqVars = {};
+          }
           outputNode.preReqVars[paramName] = {
             type: typeof paramValue,
             value: paramValue,
